% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticnet_logistic.R
\name{elasticnet_logistic}
\alias{elasticnet_logistic}
\title{Logistic regression with elasticnet}
\arguments{
\item{X_train}{Training set of explanatory variables}

\item{y_train}{Training set (vector) of explanatory variables}

\item{X_test}{Test set of explanatory variables}

\item{y_test}{Test set (vector) of explanatory variables}
}
\value{
Solution results and a range of model evaluation results \code{n}
}
\description{
The 'elasticnet_logistic' function implements a logistic regression model based on ElasticNet regularization, and selects the best ' alpha ' and ' lambda ' parameters through cross-validation. Firstly, the function selects the optimal model parameters by cross-validating different ' alpha ' values ( controlling the L1 and L2 regularization ratios ). Then, the final model is trained with the best parameters, and the prediction and evaluation are carried out on the test set, and the accuracy, precision, recall, F1 score and support are calculated. At the same time, the function also draws the ROC curve and calculates the AUC value to evaluate the classification performance of the model. Finally, the results of optimal model, model coefficient, cross validation score, test set performance index and ROC curve are returned. This process helps to optimize model performance in classification tasks and provides comprehensive performance evaluation.
}
\examples{
\dontrun{
    set.seed(42)
    X_train <- matrix(rnorm(100 * 20), ncol = 20)
    y_train <- sample(c(0, 1), 100, replace = TRUE)
    y_test <- sample(c(0, 1), 50, replace = TRUE)
    results <- elasticnet_logistic(X_train, y_train, X_test, y_test)
}
}
