% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Data preprocessing functions}
\usage{
preprocess(data)
}
\arguments{
\item{data}{Input data (data frame or matrix)}
}
\value{
standardized_data \code{n}
}
\description{
Missing value filling and near normalisation of data
}
\examples{
\dontrun{
    data <- data.frame(
        Feature1 = c(1, 2, NA, 4, 5),
        Feature2 = c(NA, 2, 3, NA, 5),
        Feature3 = c(1, 2, 3, 4, 5))
    processed_data <- preprocess(data)
    print(processed_data)
}
}
