% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stratified_train_test_split}
\alias{stratified_train_test_split}
\title{stratified_train_test_split}
\arguments{
\item{X}{A numeric matrix or data frame containing the features, where rows are samples and columns are features.}

\item{y}{A vector of class labels, where each element corresponds to the class of the sample in `X`.}

\item{test_ratio}{A numeric value between 0 and 1 specifying the proportion of the dataset to include in the test set.}
}
\value{
A list containing:
\itemize{
\item \code{X_train}: A numeric matrix or data frame of features for the training set.
\item \code{y_train}: A vector of class labels for the training set.
 \item \code{X_test}: A numeric matrix or data frame of features for the test set.
\item \code{y_test}: A vector of class labels for the test set.
}
}
\description{
Performs stratified sampling to split a dataset into training and test sets.
}
\details{
This function performs stratified sampling to split the dataset into training and test sets while ensuring that each class in `y` is proportionally represented in both the training and test sets.
}
\examples{
\dontrun{
set.seed(123)
X <- matrix(rnorm(1000), ncol = 10)  # 100 samples, 10 features
y <- sample(1:3, 100, replace = TRUE)  # 3 classes
result <- stratified_train_test_split(X, y, test_ratio = 0.2)

# Access the training and testing sets
X_train <- result$X_train
y_train <- result$y_train
X_test <- result$X_test
y_test <- result$y_test
}
}
